Filter RepoIDs from Docker to just the repository we expected
If the image we just built happens to be included in a multi-arch image
that we have locally, it seems that Docker can return that ID too. But
that doesn't work, because we need to return the arch-specific hash to
the user, so they can use it in `docker manifest`.