Allow queues to be set inactive
The command-line client now has `pause` and `unpause` commands to
control this. A paused worker will not be assigned any more items until
it is unpaused, but it will continue with any jobs it is already
running.

Also, remove the `will_cache` hashtable. We only add something to a
worker queue if it's either already cached (and `will_cache` isn't
needed), or the worker is ready and will add the item to `cache`
immediately anyway.