Improve upgrade logic
Add worker restart to scheduler. Before, the client connected directly
to the worker to ask it to restart, but then the client had no way to
know when it had reconnected.

Now, the client asks the scheduler to handle it. Also, the client can
now upgrade a whole pool at once.

Also, only try to update our Docker service when running under Docker.
Otherwise, just exit and assume the local system will restart with the
latest version.