Enable healthchecks
Check that the system is healthy on start-up (and refuse to start if
not). Then check system health before asking for a new job (but not more
than once per 10 minutes). If the system becomes unhealthy, the worker
pauses itself until it becomes healthy again. It also reports a
Prometheus metric to say it is unhealthy.