Default to GitLab merge-requests refspecs
GitLab offers self-hosted instances. It might not be possible to tell
from the URL if the repository if web-hosted on a GitLab instance.
Considering that GitHub doesn't provide self-hosted instances, default
to GitLab.

Gitea/Gogs use GitHub style, so they're currently left out.
Bitbucket doesn't seem to provide that feature.