Add logging to docker push retries
To support the logging, this reworks the `Lwt_retry` logic so that
erroneous attempts include an int representing which number of attempt
it was. This allows users of the library to easily add logging that
tracks attempt numbers, and tracks how many attempts where made leading
up to a retry exhaustion or fatal error.

We were already producing this this data structure for the sleep logic,
so this change moves that structure into the public API, since it has
general utility.